requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
eth-account>=0.9.0 